+++ ibllib/tests/test_plots.py
@@ -5,13 +5,12 @@ import uuid
 from pathlib import Path
 from PIL import Image
 from urllib.parse import urlparse
-import datetime
-import numpy as np
 
 from one.api import ONE
 from one.webclient import http_download_file
 
 from ibllib.tests import TEST_DB
+from ibllib.tests.fixtures.utils import register_new_session
 from ibllib.plots.snapshot import Snapshot
 from ibllib.plots.figures import dlc_qc_plot
 
@@ -22,7 +21,7 @@ class TestSnapshot(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        # Make a small image an store in tmp file
+        # Make a small image and store in tmp file
         cls.tmp_dir = tempfile.TemporaryDirectory()
         cls.img_file = Path(cls.tmp_dir.name).joinpath('test.png')
         image = Image.new('RGBA', size=(WIDTH, HEIGHT), color=(155, 0, 0))
@@ -34,13 +33,11 @@ class TestSnapshot(unittest.TestCase):
         cls.notes = []
 
         # make a new test session
-        date = str(datetime.date(2018, np.random.randint(1, 12), np.random.randint(1, 28)))
-        from one.registration import RegistrationClient
-        _, eid = RegistrationClient(cls.one).create_new_session('ZM_1150', date=date)
+        _, eid = register_new_session(cls.one, subject='ZM_1150')
         cls.eid = str(eid)
 
     def _get_image(self, url):
-        # This is a bit of a hack because when running a the server locally, the request to the media folder fail
+        # This is a bit of a hack because when running the server locally, the request to the media folder fails
         rel_path = urlparse(url).path[1:]
         try:
             img_file = list(Path('/var/www/').rglob(rel_path))[0]